requirements.txt
torch
torchvision
transformers
scipy
scikit-learn
pillow
pyyaml
opencv-python
lightglue
